SafelyWritableFile class moved to utility folder.
