+++ CMakeLists.txt
@@ -928,6 +928,7 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/util/color/colorpalette.cpp
   src/util/color/predefinedcolorpalettes.cpp
   src/util/console.cpp
+  src/util/safelywritablefile.cpp
   src/util/db/dbconnection.cpp
   src/util/db/dbconnectionpool.cpp
   src/util/db/dbconnectionpooled.cpp
